+++ RGBeta/RGBeta.m
@@ -1,7 +1,7 @@
 (*
 ################################################################
-This is RGBeta, Version 1.1.1.
-Last modefied 17-01-2022
+This is RGBeta, Version 1.1.2.
+Last modefied 18-05-2022
 
 Released under the MIT license (see 'LICENSE').
 
@@ -127,7 +127,7 @@ Protect[$a, $b, $c, $d, $i, $j, $A, $B];
 Protect[SO, Sp, SU, U1];
 Protect[f1, f2, s1, s2, s3, s4];
 
-$RGBetaVersion= "v1.1.1"
+$RGBetaVersion= "v1.1.2"
 
 
 (*###########################################*)